move flag alignments to common
